feat(l_odyssee): implement rudimentary deck generation
Took a bit of work, but deck generation went by smoothly. Had to
update all of the type stubs to genanki to get it to work. Considering
just forking genanki and doing it myself.

Pro8a8ly broke something. This early on, I'm not going to worry about
breaking changes lmao.