fix(genanki_ext)!: set guid in Note subclass
Previously, `guid` was set to an empty string. Apparently, this is a
Very Bad Idea. Anki didn't recognize a hundred or so cards of mine. The
solution is: set `guid` in the way that the documentation seems to think
that it should be set.

See here for some more information:
  https://docs.ankiweb.net/importing/text-files.html?highlight=guid#guid-column

BREAKING CHANGES: L'Odyssée is broken again! The Merge Duplicate Notes
plugin is, again, a lifesaver: https://ankiweb.net/shared/info/55394168